Token counting support (approximate) added
An input token count is now printed before each completion request. This is to make it easier to gauge how much is being spent on the conversation, and to keep track of context length until a solution is implemented to drop excessive chat log messages in a smart way (will need a refactor, I don't have time to implement this at this time.